option to download planet file via url
